Refactor AmpContext dependencies for building (#8258)
* Fix types for AmpContext

* Move and split dependencies for ampcontext-lib

* Enable type checks for ampcontext-lib
