+++ 3p/ampcontext.js
@@ -13,24 +13,44 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import './polyfills';
 import {dev} from '../src/log';
 import {IframeMessagingClient} from './iframe-messaging-client';
-import {MessageType} from '../src/3p-frame';
+import {MessageType} from '../src/3p-frame-messaging';
 import {tryParseJson} from '../src/json';
 
 export class AmpContext {
 
   /**
-   *  @param {Window} win The window that the instance is built inside.
+   *  @param {!Window} win The window that the instance is built inside.
    */
   constructor(win) {
+    /** @private {!Window} */
     this.win_ = win;
+
+    /** @type {?string} */
+    this.location = null;
+
+    /** @type {?string} */
+    this.canonicalUrl = null;
+
+    /** @type {?string} */
+    this.pageViewId = null;
+
+    /** @type {?string} */
+    this.sentinel = null;
+
+    // TODO(alanorozco): confirm type
+    /** @type {?} */
+    this.startTime = null;
+
+    // TODO(alanorozco): confirm type
+    /** @type {?string} */
+    this.referrer = null;
 
     this.findAndSetMetadata_();
     this.client_ = new IframeMessagingClient(win);
     this.client_.setHostWindow(this.getHostWindow_());
-    this.client_.setSentinel(this.sentinel);
+    this.client_.setSentinel(dev().assertString(this.sentinel));
   }
 
   /**
@@ -64,8 +84,8 @@ export class AmpContext {
   /**
    *  Send message to runtime requesting to resize ad to height and width.
    *    This is not guaranteed to succeed. All this does is make the request.
-   *  @param {int} width The new width for the ad we are requesting.
-   *  @param {int} height The new height for the ad we are requesting.
+   *  @param {number} width The new width for the ad we are requesting.
+   *  @param {number} height The new height for the ad we are requesting.
    */
   requestResize(width, height) {
     this.client_.sendMessage(MessageType.EMBED_SIZE, {width, height});
@@ -75,8 +95,8 @@ export class AmpContext {
    *  Allows a creative to set the callback function for when the resize
    *    request returns a success. The callback should be set before resizeAd
    *    is ever called.
-   *  @param {function(requestedHeight, requestedWidth)} callback Function
-   *    to call if the resize request succeeds.
+   *  @param {function(number, number)} callback Function to call if the resize
+   *    request succeeds.
    */
   onResizeSuccess(callback) {
     this.client_.registerCallback(MessageType.EMBED_SIZE_CHANGED, obj => {
@@ -87,8 +107,8 @@ export class AmpContext {
    *  Allows a creative to set the callback function for when the resize
    *    request is denied. The callback should be set before resizeAd
    *    is ever called.
-   *  @param {function(requestedHeight, requestedWidth)} callback Function
-   *    to call if the resize request is denied.
+   *  @param {function(number, number)} callback Function to call if the resize
+   *    request is denied.
    */
   onResizeDenied(callback) {
     this.client_.registerCallback(MessageType.EMBED_SIZE_DENIED, obj => {
@@ -99,7 +119,7 @@ export class AmpContext {
   /**
    *  Takes the current name on the window, and attaches it to
    *  the name of the iframe.
-   *  @param {HTMLIframeElement} iframe The iframe we are adding the context to.
+   *  @param {HTMLIFrameElement} iframe The iframe we are adding the context to.
    */
   addContextToIframe(iframe) {
     iframe.name = this.win_.name;
@@ -108,15 +128,15 @@ export class AmpContext {
   /**
    *  Parse the metadata attributes from the name and add them to
    *  the class instance.
-   *  @param {!Object|string} contextData
+   *  @param {!Object|string} data
    *  @private
    */
   setupMetadata_(data) {
-    data = tryParseJson(data);
-    if (!data) {
+    const dataObject = typeof data === 'string' ? tryParseJson(data) : data;
+    if (!dataObject) {
       throw new Error('Could not setup metadata.');
     }
-    const context = data._context;
+    const context = dataObject._context;
     this.location = context.location;
     this.canonicalUrl = context.canonicalUrl;
     this.pageViewId = context.pageViewId;
@@ -149,6 +169,8 @@ export class AmpContext {
   findAndSetMetadata_() {
     // If the context data is set on window, that means we don't need
     // to check the name attribute as it has been bypassed.
+    // TODO(alanorozco): why the heck could AMP_CONTEXT_DATA be two different
+    // types? FIX THIS.
     if (!this.win_.AMP_CONTEXT_DATA) {
       this.setupMetadata_(this.win_.name);
     } else if (typeof this.win_.AMP_CONTEXT_DATA == 'string') {
