+++ src/event-helper-listen.js
@@ -1,0 +1,54 @@
+/**
+ * Copyright 2017 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * Listens for the specified event on the element.
+ *
+ * Do not use this directly. This method is implemented as a shared
+ * dependency. Use `listen()` in either `event-helper` or `3p-frame-messaging`,
+ * depending on your use case.
+ *
+ * @param {!EventTarget} element
+ * @param {string} eventType
+ * @param {function(!Event)} listener
+ * @param {boolean=} opt_capture
+ * @return {!UnlistenDef}
+ */
+export function internalListenImplementation(element, eventType, listener,
+    opt_capture) {
+  let localElement = element;
+  let localListener = listener;
+  /** @type {?Function}  */
+  let wrapped = event => {
+    try {
+      return localListener.call(this, event);
+    } catch (e) {
+      // reportError is installed globally per window in the entry point.
+      self.reportError(e);
+      throw e;
+    }
+  };
+  const capture = opt_capture || false;
+  localElement.addEventListener(eventType, wrapped, capture);
+  return () => {
+    if (localElement) {
+      localElement.removeEventListener(eventType, wrapped, capture);
+    }
+    localListener = null;
+    localElement = null;
+    wrapped = null;
+  };
+}
