+++ test/functional/3p/test-iframe-messaging-client.js
@@ -15,7 +15,7 @@
  */
 
 import {IframeMessagingClient} from '../../../3p/iframe-messaging-client';
-import {serializeMessage} from '../../../src/3p-frame';
+import {serializeMessage} from '../../../src/3p-frame-messaging';
 
 describes.realWin('iframe-messaging-client', {}, env => {
